Move libraryUtils to test folder
