+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@setprotocol/set-protocol-v2",
-  "version": "0.0.29",
+  "version": "0.0.30",
   "description": "",
   "main": "dist",
   "types": "dist/types",
