+++ utils/common/index.ts
@@ -34,6 +34,3 @@ export {
   calculateTokensInReserve,
   getReservesSafe
 } from "./uniswapUtils";
-export {
-  convertLibraryNameToLinkId
-} from "./libraryUtils";
